move logger to internal
