+++ internal/provider/provider.go
@@ -7,9 +7,8 @@ import (
 	"context"
 	"os"
 
+	"github.com/onmetal/inventory/internal/logger"
 	"github.com/urfave/cli/v2"
-
-	"github.com/onmetal/inventory/cmd/benchmark-scheduler/logger"
 )
 
 const (
