+++ internal/smbiosinfo/smbios.go
@@ -12,9 +12,8 @@ import (
 	_ "unsafe"
 
 	"github.com/digitalocean/go-smbios/smbios"
+	"github.com/onmetal/inventory/internal/logger"
 	"github.com/pkg/errors"
-
-	"github.com/onmetal/inventory/cmd/benchmark-scheduler/logger"
 )
 
 const (
